SD intro updates
This is mainly changing references to other sections to be dynamic links. The purpose here is to let asciidoctor automatically generate the section numbers (a bonus is internal cross-linking in the output), so if a new section is added, it will be automatically renumbered without needing to review the document explicitly.