SD FCS Updates
Using FCS_RBG_EXT.2 & FCS_STO_EXT.1 as a comparison, they provide different methods for presenting similar sets of tests. In each case there are tests specific to Windows/Linux/macOS, but they are presented differently in each EA.

For the RBG, there are 3 tests, each listed as conditional, with one test for each platform.

For the STO, there is a single test, and then descriptions as to how to test on each of the 3 platforms.

The presentation of the testing should be done consistently throughout all the EAs.

The updates here to FCS_STO_EXT.1 are to match it to the rest of the tests in the SD.